Move CoreCaseDataApiMockHelper to test package
