+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/scheduler/JudgementBufferSchedulerITest.java
@@ -17,6 +17,7 @@ import uk.gov.hmcts.reform.civil.scheduler.judgementbuffer.JudgementBufferSchedu
 import uk.gov.hmcts.reform.civil.service.FeatureToggleService;
 import uk.gov.hmcts.reform.civil.service.TelemetryService;
 import uk.gov.hmcts.test.config.CoreCaseDataApiMockHelperConfiguration;
+import uk.gov.hmcts.test.helper.CoreCaseDataApiMockHelper;
 
 import java.util.List;
